refactor user domain imports for consistency and clarity
