+++ src/main/java/depromeet/lessonfour/server/auth/api/controller/AuthTestController.java
@@ -22,6 +22,6 @@ public class AuthTestController {
 
   @GetMapping("/{email}")
   public ResponseEntity<?> getUserByEmail(@PathVariable String email) {
-    return ResponseEntity.ok(userQueryService.getUserForAuthentication(email));
+    return ResponseEntity.ok(userQueryService.getUserByEmail(email));
   }
 }
