+++ src/main/java/depromeet/lessonfour/server/auth/app/service/KakaoAuthService.java
@@ -24,10 +24,10 @@ import depromeet.lessonfour.server.auth.infra.security.jwt.JwtTokenGenerator;
 import depromeet.lessonfour.server.common.exception.ServerException;
 import depromeet.lessonfour.server.user.app.dto.response.AuthResponseDto;
 import depromeet.lessonfour.server.user.app.dto.response.UserResponseDto;
-import depromeet.lessonfour.server.user.domain.Provider;
-import depromeet.lessonfour.server.user.domain.Provider.ProviderType;
-import depromeet.lessonfour.server.user.domain.User;
-import depromeet.lessonfour.server.user.domain.UserRepository;
+import depromeet.lessonfour.server.user.domain.entity.User;
+import depromeet.lessonfour.server.user.domain.repository.UserRepository;
+import depromeet.lessonfour.server.user.domain.vo.Provider;
+import depromeet.lessonfour.server.user.domain.vo.Provider.ProviderType;
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.Locator;
