+++ src/main/java/depromeet/lessonfour/server/auth/app/service/RefreshTokenUseCase.java
@@ -8,8 +8,8 @@ import depromeet.lessonfour.server.auth.domain.vo.AccountContext;
 import depromeet.lessonfour.server.auth.infra.security.jwt.JwtTokenGenerator;
 import depromeet.lessonfour.server.auth.infra.security.jwt.JwtTokenValidator;
 import depromeet.lessonfour.server.common.annotation.UseCase;
-import depromeet.lessonfour.server.user.domain.User;
-import depromeet.lessonfour.server.user.domain.UserRepository;
+import depromeet.lessonfour.server.user.domain.entity.User;
+import depromeet.lessonfour.server.user.domain.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 
 @UseCase
