+++ src/main/java/depromeet/lessonfour/server/user/app/dto/response/UserProfileResponseDto.java
@@ -1,6 +1,6 @@
 package depromeet.lessonfour.server.user.app.dto.response;
 
-import depromeet.lessonfour.server.user.domain.User;
+import depromeet.lessonfour.server.user.domain.entity.User;
 
 public record UserProfileResponseDto(Long id, String email, String nickname, String profileImage) {
