+++ src/main/java/depromeet/lessonfour/server/user/app/dto/response/UserResponseDto.java
@@ -1,7 +1,7 @@
 package depromeet.lessonfour.server.user.app.dto.response;
 
-import depromeet.lessonfour.server.user.domain.Provider;
-import depromeet.lessonfour.server.user.domain.User;
+import depromeet.lessonfour.server.user.domain.entity.User;
+import depromeet.lessonfour.server.user.domain.vo.Provider;
 
 public record UserResponseDto(
     Long id, String email, String nickname, Provider provider, String profileImage, boolean isNew) {
