+++ src/main/java/depromeet/lessonfour/server/user/app/service/UserUpdateService.java
@@ -3,7 +3,7 @@ package depromeet.lessonfour.server.user.app.service;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import depromeet.lessonfour.server.user.domain.UserRepository;
+import depromeet.lessonfour.server.user.domain.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 
 @Service
