+++ src/main/java/depromeet/lessonfour/server/user/infra/repository/JpaUserRepository.java
@@ -4,7 +4,7 @@ import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import depromeet.lessonfour.server.user.domain.User;
+import depromeet.lessonfour.server.user.domain.entity.User;
 
 public interface JpaUserRepository extends JpaRepository<User, Long> {
