+++ src/test/java/depromeet/lessonfour/server/auth/security/jwt/JwtTokenGeneratorTest.java
@@ -19,7 +19,7 @@ import org.springframework.test.util.ReflectionTestUtils;
 import depromeet.lessonfour.server.auth.domain.vo.AccountContext;
 import depromeet.lessonfour.server.auth.infra.security.jwt.JwtSecretKeyProvider;
 import depromeet.lessonfour.server.auth.infra.security.jwt.JwtTokenGenerator;
-import depromeet.lessonfour.server.user.domain.User;
+import depromeet.lessonfour.server.user.domain.entity.User;
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.security.Keys;
